+++ include/GafferScene/ScenePath.h
@@ -76,6 +76,7 @@ class GAFFERSCENE_API ScenePath : public Gaffer::Path
 		void setContext( Gaffer::ContextPtr context );
 		Gaffer::Context *getContext();
 		const Gaffer::Context *getContext() const;
+		const Gaffer::Context *inspectionContext( const IECore::Canceller *canceller = nullptr ) const;
 
 		bool isValid( const IECore::Canceller *canceller = nullptr ) const override;
 		bool isLeaf( const IECore::Canceller *canceller = nullptr ) const override;